Extract graph helpers from App
